+++ dev/main-react16/src/pages/react16/react16.js
@@ -4,7 +4,7 @@ import jsxCustomEvent from '@micro-zoe/micro-app/polyfill/jsx-custom-event'
 import React from 'react'
 import { Spin, Row, Col, Button, Modal } from 'antd'
 import { LoadingOutlined } from '@ant-design/icons';
-import microApp from '@micro-zoe/micro-app'
+import microApp, { unmountApp, unmountAllApps } from '@micro-zoe/micro-app'
 import config from '../../config'
 import './react16.less'
 
@@ -112,6 +112,16 @@ export default class App extends React.Component {
     })
   }
 
+  // 主动卸载应用
+  handleUnmountMySelf = () => {
+    unmountApp(this.state.name, {
+      // destroy: true,
+      // clearAliveState: true,
+    }).then(() => {
+      console.log('unmountApp方法 -- 卸载成功')
+    })
+  }
+
   changeTestNum = () => {
     this.setState({
       testNum: this.state.testNum + 1,
@@ -149,6 +159,7 @@ export default class App extends React.Component {
             <Button type="primary" onClick={this.dispatchGlobalData}>发送全局数据</Button>
             <Button type="primary" onClick={this.changeNameUrl}>切换应用</Button>
             <Button type="primary" onClick={this.handleModal}>modal内嵌应用</Button>
+            <Button type="primary" onClick={this.handleUnmountMySelf}>主动卸载应用</Button>
           </Col>
           <Col span={18} className='app-con-react16'>
             { this.state.showLoading && <Spin indicator={antIcon} /> }
@@ -173,7 +184,9 @@ export default class App extends React.Component {
                   // destroy
                   // inline
                   // disableSandbox
+                  // disable-sandbox
                   // disableScopecss
+                  // disable-scopecss
                   // macro
                 >
                 </micro-app>